% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_normalization.R
\name{counts_normalization}
\alias{counts_normalization}
\title{Calulate CPM Normalization for Raw Gene Expression Counts}
\usage{
counts_normalization(
  raw_counts,
  MARGIN = c("row", "col"),
  log2 = TRUE,
  plot = TRUE,
  TMM_method = FALSE
)
}
\arguments{
\item{raw_counts}{numeric matrix or data frame of raw gene expression counts.}

\item{MARGIN}{"row" or "col" to identify the place of sample (library size). If the \code{raw_counts} matrix has a \code{nxG} size, put "row" if it's opposite put "col".}

\item{log2}{a logical indicating whether the normalized counts will be transformed in log2. Default is \code{TRUE}.}

\item{plot}{a logical indicating whether a graph will be displayed for the normalization check. Default is \code{TRUE}.}

\item{TMM_method}{a logical indicating whether the library size will be normalized by TMM method of \code{edgeR} package. Default is \code{FALSE}.}
}
\value{
normalized counts matrix
}
\description{
Calulate CPM Normalization for Raw Gene Expression Counts
}
\examples{

#Generate data
sim_Genes <- c(runif(90,0,1), rbinom(60,2,0.4), runif(30,1250,2000), runif(20,100,300))
raw_data <- matrix(sim_Genes, nrow = 10)
colnames(raw_data) <- paste0("G", sample(1:ncol(raw_data), ncol(raw_data), replace = FALSE))
rownames(raw_data) <- paste0("Sample_", 1:nrow(raw_data))

#Use the method
norm_data <- counts_normalization(raw_counts = data.frame(raw_data), MARGIN = "row")
}
\author{
Mélanie Huchon
}
