% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMngt_piechart.R
\name{dataMngt_piechart}
\alias{dataMngt_piechart}
\title{Function to manage data for the piecharts of polyfunctionnality of ICS data}
\usage{
dataMngt_piechart(
  data,
  list_cyto,
  ID,
  group,
  group_val = NULL,
  Timepoint,
  Timepoint_val = NULL,
  Stim,
  Stim_val,
  pop_type,
  gp_cyto = NULL
)
}
\arguments{
\item{data}{}

\item{list_cyto}{}

\item{ID}{a character to define the colnames of ID in data_ICS}

\item{group}{a character to define the colnames of group in data_ICS (Arm)}

\item{group_val}{default is NULL : no filter}

\item{Timepoint}{a character to define the colnames of Time point in data_ICS}

\item{Timepoint_val}{default is NULL : no filter}

\item{Stim}{a character to define the colnames of stimulation in data_ICS}

\item{Stim_val}{}

\item{pop_type}{a character to define the type of population (CD4 or CD8) (if your cytokines are for example : "IFN+IL2+MIP1B+ in CD4" in data)}

\item{gp_cyto}{vector with the number of cytokines that you want to group (ex : c(2,3))}
}
\value{
a \code{data.frame} object
}
\description{
The boolean cytokines must have this type of format : IFN-IL2-MIP1b-TNF+ in CD4
}
\examples{

\dontrun{d <- dataMngt_piechart(data = ICS_data, list_cyto = c("IFN", "IL2", "MIP1b", "TNF"),
ID = "ID", group = "Arm", group_val = "Arm1", Timepoint = "TP", Timepoint_val = "TP1",
Stim = "Stim", Stim_val = "BackgroundSubstracted", pop_type = "CD4")}
}
\author{
Mélanie Huchon
}
\keyword{internal}
