% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_table.R
\name{desc_table}
\alias{desc_table}
\title{Descriptive table function}
\usage{
desc_table(data, vars)
}
\arguments{
\item{data}{}

\item{variables}{}
}
\value{

}
\description{
Descriptive table function
}
\examples{

}
