% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pie_arcs.R
\name{draw_pie_arcs}
\alias{draw_pie_arcs}
\title{Draw pie charts with arcs}
\usage{
draw_pie_arcs(
  value,
  arcs,
  group,
  title,
  leg = F,
  size,
  piecolors = NULL,
  ringcolors = NULL
)
}
\arguments{
\item{value}{}

\item{arcs}{}

\item{group}{}

\item{title}{}

\item{leg}{}

\item{size}{}

\item{pie_col}{}

\item{ring_col}{}
}
\value{
a \code{pie} object
}
\description{
This function is to do plots for polyfunctionnality of ICS data
}
\author{
Claire Bauduin (adapted to Binbin Xu codes) - Last modifications by Mélanie Huchon
}
\keyword{internal}
