% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fingerprint.R
\name{Fingerprint}
\alias{Fingerprint}
\title{Function for Finger print plot on Gene set analysis}
\usage{
Fingerprint(
  res_dgsaseq,
  datalog2FC,
  genesets = "Chaussabel",
  category = "immune",
  time = NULL,
  alpha = 0.05,
  title = NULL,
  max_size = 8,
  size_y_lab = 9,
  order_by_pvals = FALSE,
  interactive.plot = FALSE
)
}
\arguments{
\item{res_dgsaseq}{an object of \code{dgsa_seq} function in \code{dearseq} package.}

\item{datalog2FC}{a data.frame with genes in rownames and if (time == NULL) 1 column with Fold Change else several columns are allowed but colnames must contain time's string characters.}

\item{genesets}{a string character "BTM" or "Chaussabel" to use genesets included or an other gmt object. Default is \code{Chaussabel}.}

\item{category}{a string character or vector. Only needed if \code{(genesets == "BTM")} : "immune", "signaling", "TF targets", "biological process", "molecular function", "location", "all".}

\item{time}{a string character to filter on time. This time must be present on \code{datalog2FC}'s colnames. Default is \code{NULL}.}

\item{alpha}{a numeric corresponding to type 1 error threshold. Default value is \code{0.05}.}

\item{title}{a string character for plot's title. Default is \code{NULL}.}

\item{max_size}{a numeric corresponding to the maximum size of point. Default is \code{8}.}

\item{size_y_lab}{labsize for non-interactive plot, can be useful to set to 0 to combine several plots}

\item{order_by_pvals}{a logical indicating to order genesets per pvalue left to right. Default is \code{FALSE}.}

\item{interactive.plot}{a logical indicating whether the plot will be interactive, recommended if order_by_pvals is \code{TRUE}. Default is \code{FALSE}.}
}
\value{
a \code{ggplot} object
}
\description{
The aim of this function is to create a visual representation of gene set enrichment results from differential gene expression analysis.
}
\author{
Mélanie Huchon & Quentin Laval
}
