% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleBoxplot.R
\name{multipleBoxplots}
\alias{multipleBoxplots}
\title{Multiple boxplots for many times}
\usage{
multipleBoxplots(data, x_var, y_var)
}
\arguments{
\item{data}{a dataset from which the variable \code{x_var} and \code{y_var} should be taken.}

\item{x_var}{corresponding to the x coordinates for the plot.}

\item{y_var}{corresponding to the y coordinates for the plot.}
}
\value{

}
\description{
Multiple boxplots for many times
}
\examples{
data <- read.table("C:/Users/mh8/Downloads/data_pkg.txt")
multipleBoxplots( data = data, x_var = time, y_var = TNFRSF9)
}
