% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol2cart_piechart.R
\name{pol2cart}
\alias{pol2cart}
\title{Convert polar coordinates to Cartesian coordinates}
\usage{
pol2cart(theta, r, deg = FALSE, recycle = FALSE)
}
\arguments{
\item{theta}{}

\item{r}{}

\item{deg}{}

\item{recycle}{}
}
\value{
a \code{matrix} object
}
\description{
Convert polar coordinates to Cartesian coordinates
}
\author{
Claire Bauduin (adapted to Binbin Xu codes) - Last modifications by Mélanie Huchon
}
\keyword{internal}
