% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVCA.R
\name{PVCA}
\alias{PVCA}
\title{Performs a Principal Variance Component Analysis (PVCA) and output the plot}
\usage{
PVCA(
  expression,
  sample_info,
  Sample_ID,
  effects,
  pct_threshold = 0.95,
  verbose = FALSE,
  title = "PVCA",
  ...
)
}
\arguments{
\item{expression}{numeric matrix of gene expressions with sample IDs as \code{colnames} and gene/probes in rows}

\item{sample_info}{a data.frame containing the design and technical informations on the samples.}

\item{Sample_ID}{the name of the column in \code{sample_info} which contain the sample IDs and which should match \code{colnames(expression)} in the same order.}

\item{effects}{the name of the columns from \code{sample_info} which should be investigated for potential Batch effect}

\item{pct_threshold}{a value between 0 and 1 (Default is \code{0.95}). The percentile value of the minimum amount of the variabilities that the selected principal components need to explain.}

\item{verbose}{a logical indicating if information about method's progress must be printed. Default is \code{FALSE}.}

\item{title}{the title of the barplot.}

\item{...}{others parameters to be passed into the \code{barplot}.}
}
\value{
a barplot.
}
\description{
Performs a Principal Variance Component Analysis (PVCA) and output the plot
}
\references{
Adapted from Pierre R. Bushel original programm 
(email: \code{Bushel@niehs.nih.gov})
}
\author{
Boris Hejblum
}
