% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sistmr-package.r
\docType{package}
\name{sistmr-package}
\alias{sistmr}
\alias{sistmr-package}
\title{sistmr: A Collection of Utility Functions from the Inserm/Inria SISTM Team}
\description{
Functions common to members of the SISTM team.
}
\keyword{internal}
