% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_GS.R
\name{table_GS}
\alias{table_GS}
\title{General FingerPrintplot, for BTM and Chaussabel modules}
\usage{
table_GS(res_dgsaseq, data_log2FC, time = NULL, geneset_modules)
}
\arguments{
\item{res_dgsaseq}{an object of \code{dgsa_seq} function in \code{dearseq} package.}

\item{data_log2FC}{a data.frame with genes in rownames and if (time == NULL) 1 column with Fold Change else several columns are allowed but colnames must contain time's string characters.}

\item{time}{a string character to filter on time. This time must be present on \code{datalog2FC}'s colnames. Default is \code{NULL}.}

\item{geneset_modules}{a string character "BTM" or "Chaussabel" to use genesets included or an other gmt object. Default is \code{Chaussabel}.}
}
\value{
a \code{data.frame} object
}
\description{
General FingerPrintplot, for BTM and Chaussabel modules
}
\keyword{internal}
