% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleBoxplot.R
\name{multipleBoxplots}
\alias{multipleBoxplots}
\title{Multiple boxplots for many times}
\usage{
multipleBoxplots(abs, ord, data, color_palette = "Dark2", color_manual = FALSE)
}
\arguments{
\item{abs}{corresponding to the x coordinates for the plot.}

\item{ord}{corresponding to the y coordinates for the plot.}

\item{data}{Dataset to use for plot.}

\item{color_palette}{a color palette for boxplot.}

\item{color_manual}{a color vector where each color is for one boxplot.}
}
\value{

}
\description{
Multiple boxplots for many times
}
